Reorganization of code
