+++ src/cpm/copy_pointer_manager.cpp
@@ -1,0 +1,98 @@
+#include <fstream>
+#include <algorithm>
+#include <list>
+#include "copy_pointer_manager.hpp"
+
+bool SimpleCopyPointerManager::registerCopyPointer(std::string pattern, size_t position) {
+    if (pointer_map.count(pattern) == 0) {
+
+        struct SimplePointerInfo pattern_info = {
+            .pointers = {position},
+            .copy_pointer_index = 0,
+        };
+
+        pointer_map.insert({pattern, pattern_info});
+        
+        return false;
+    }
+    
+    pointer_map[pattern].pointers.push_back(position);
+    return true;
+}
+
+int SimpleCopyPointerManager::getCopyPointer(std::string pattern) {
+    return pointer_map[pattern].pointers[pointer_map[pattern].copy_pointer_index];
+}
+
+void SimpleCopyPointerManager::reportPrediction(std::string pattern, bool hit) {
+    if (hit) {
+        hits++;
+    } else {
+        misses++;
+    }
+}
+
+void SimpleCopyPointerManager::reset() {
+    hits = 0;
+    misses = 0;
+}
+
+int SimpleCopyPointerManager::getHits(std::string current_pattern) { return hits; }
+
+int SimpleCopyPointerManager::getMisses(std::string current_pattern) { return misses; }
+
+void RecentCopyPointerManager::repositionCopyPointer(std::string pattern, ReadingStrategy* reading_strategy) {
+    // second to last copy pointer (because most recent could lead to predicting future)
+    pointer_map[pattern].copy_pointer_index = pointer_map[pattern].pointers.size() - 2;
+}
+
+void NextOldestCopyPointerManager::repositionCopyPointer(std::string pattern, ReadingStrategy* reading_strategy) {
+    pointer_map[pattern].copy_pointer_index += 1;
+}
+
+void MostCommonCopyPointerManager::repositionCopyPointer(std::string pattern, ReadingStrategy* reading_strategy) {
+    
+    std::list<size_t> pointer_candidates(pointer_map[pattern].pointers.begin(), pointer_map[pattern].pointers.end());
+    
+    int count;
+    int offset = 1;
+    char most_frequent = '\0';
+
+    while (most_frequent == '\0' || count > 1){
+
+        count = 0;
+
+        // Majority algorithm: first pass (determine most frequent)
+        for (size_t pointer : pointer_candidates) {
+            char char_at_pointer = reading_strategy->at(pointer + offset);
+
+            if (count == 0) {
+                most_frequent = char_at_pointer;
+                count++;
+            } else if (most_frequent == char_at_pointer) {
+                count++;
+            } else {
+                count--;
+            }
+        }
+
+        // Majority algorithm: second pass (remove all pointers that don't match the most frequent)
+        for (std::list<size_t>::iterator it = pointer_candidates.begin(); it != pointer_candidates.end();) {
+            size_t pointer = *it;
+            if (reading_strategy->at(pointer + offset) == most_frequent)
+                it++;
+            else
+                it = pointer_candidates.erase(it);
+        }
+
+        offset++;
+    }
+
+    size_t pointer_candidate = pointer_candidates.front();
+    int i;
+    for (i = 0; i < pointer_map[pattern].pointers.size(); i++)
+        if (pointer_map[pattern].pointers[i] == pointer_candidate)
+            break;
+    pointer_map[pattern].copy_pointer_index = i;
+
+}
