+++ src/cpm/cpm.hpp
@@ -2,111 +2,9 @@
 #include <string>
 #include <map>
 #include "parser.hpp"
-
-
-class CopyPointerThreshold {
-public:
-    virtual bool surpassedThreshold(double hit_probability) = 0;
-    virtual void reset() = 0;
-};
-
-class StaticCopyPointerThreshold : public CopyPointerThreshold {
-
-    double static_threshold;
-
-public:
-    StaticCopyPointerThreshold(double st) : static_threshold(st) {}
-    bool surpassedThreshold(double hit_probability);
-    void reset();
-};
-
-class DerivativeCopyPointerThreshold : public CopyPointerThreshold {
-
-    double previous_hit_probability = -1;
-    double derivative_threshold;
-
-public:
-    DerivativeCopyPointerThreshold(double dt) : derivative_threshold(dt) {}
-    bool surpassedThreshold(double hit_probability);
-    void reset();
-};
-
-class SuccessFailsCopyPointerThreshold : public CopyPointerThreshold {
-
-    double previous_hit_probability = -1;
-    int fails_count = 0;
-    int fails_threshold;
-
-public:
-    SuccessFailsCopyPointerThreshold(double dt) : fails_threshold(dt) {}
-    bool surpassedThreshold(double hit_probability);
-    void reset();
-};
-
-
-struct SimplePointerInfo {
-    std::vector<size_t> pointers;
-    int copy_pointer_index;     // Pattern's last symbol
-};
-
-class CopyPointerManager {
-public:
-    virtual int getCopyPointer(std::string) = 0;
-    virtual void repositionCopyPointer(std::string, ReadingStrategy*) = 0;
-    virtual bool registerCopyPointer(std::string, size_t) = 0;
-    virtual void reportPrediction(std::string, bool) = 0;
-    virtual void reset() = 0;
-    virtual int getHits(std::string) = 0;
-    virtual int getMisses(std::string) = 0;
-};
-
-class SimpleCopyPointerManager : public CopyPointerManager {
-
-protected:
-    std::map<std::string, SimplePointerInfo> pointer_map;
-    int hits = 0;
-    int misses = 0;
-
-public:
-    int getCopyPointer(std::string);
-    bool registerCopyPointer(std::string, size_t);
-    void reportPrediction(std::string, bool);
-    void reset();
-    int getHits(std::string);
-    int getMisses(std::string);
-};
-
-class MostCommonCopyPointerManager : public SimpleCopyPointerManager {
-public:
-    void repositionCopyPointer(std::string, ReadingStrategy*);    
-};
-
-class RecentCopyPointerManager : public SimpleCopyPointerManager {
-public:
-    void repositionCopyPointer(std::string, ReadingStrategy*);
-};
-
-class NextOldestCopyPointerManager : public SimpleCopyPointerManager {
-public:
-    void repositionCopyPointer(std::string, ReadingStrategy*);
-};
-
-
-class BaseDistribution {
-public:
-    virtual void setBaseDistribution(std::map<char, int> histogram) = 0;
-    std::map<char, double> distribution;
-};
-
-class UniformDistribution : public BaseDistribution {
-public:
-    void setBaseDistribution(std::map<char, int> histogram);
-};
-
-class FrequencyDistribution : public BaseDistribution {
-public:
-    void setBaseDistribution(std::map<char, int> histogram);
-};
+#include "copy_pointer_threshold.hpp"
+#include "copy_pointer_manager.hpp"
+#include "base_distribution.hpp"
 
 
 class CopyModel {
